requirements.txt
fastapi
uvicorn[standard]
boto3
requests
openai
pydantic
sqlalchemy
jwt
pyjwt
pyjwt[crypto]
psycopg2-binary